병합본
고칠 점
1. 점 클릭 시 사라지는 기능
2. 피버모드 배경 위치 조절

+++ UICombineVer1.0.0/Assets/Script/Score.cs
@@ -5,7 +5,7 @@ using UnityEngine.UI;
 
 public class Score : MonoBehaviour
 {
-    public static int score = 0;
+    public static float score = 0;
     public static int previousScore = 0;
     //GameObject DotGenerator = GameObject.Find("DotGenerator");
     // Start is called before the first frame update
